+++ sail/provision.py
@@ -1,4 +1,4 @@
-from sail import cli, util, deploy, __version__
+from sail import cli, util, deploy, domains, __version__
 
 import sail
 import json
@@ -8,6 +8,7 @@ import shutil
 import digitalocean
 import re, secrets, string
 import shlex, time, io, pathlib, jinja2
+import packaging.version
 
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives import serialization
@@ -19,9 +20,11 @@ from prettytable import PrettyTable
 @click.option('--size', default='s-1vcpu-1gb-intel', help='The Droplet size, defaults to s-1vcpu-1gb-intel. To get a full list of available sizes and pricing run: sail sizes')
 @click.option('--region', default='ams3', help='The region to deploy to, defaults to ams3. For a full list of available regions run: sail regions')
 @click.option('--email', help='The admin e-mail address. You can set the default e-mail address with: sail config email <email>')
+@click.option('--namespace', default='default', help='The namespace to use for the new application')
+@click.option('--environment', help='Initialize the application into an existing environment')
 @click.option('--force', '-f', is_flag=True)
 @click.pass_context
-def init(ctx, provider_token, email, size, region, force):
+def init(ctx, provider_token, email, size, region, force, namespace, environment):
 	'''Initialize and provision a new project'''
 	root = util.find_root()
 
@@ -31,6 +34,37 @@ def init(ctx, provider_token, email, size, region, force):
 	files = os.listdir(path='.')
 	if files and not force:
 		raise click.ClickException('This project directory is not empty, can not init here. Override with --force, can destroy existing files.')
+
+	if namespace and not re.match(r'^[a-zA-Z0-9_.-]+$', namespace):
+		raise click.ClickException('Invalid namespace. Namespaces can be alpha-numeric and contain the following characters: _.-')
+
+	namespace = namespace.lower()
+
+	if environment:
+		environment = pathlib.Path(environment) / '.sail'
+		if not environment.is_dir():
+			raise click.ClickException('Not a valid Sail environment: %s' % environment.resolve())
+
+		if not (environment / 'config.json').is_file():
+			raise click.ClickException('Not a valid Sail environment: %s' % environment.resolve())
+
+		if not (environment / 'ssh.key').is_file() or not (environment / 'ssh.key.pub').is_file():
+			raise click.ClickException('Could not read SSH keys from: %s' % environment.resolve())
+
+		with (environment / 'config.json').open('r') as f:
+			env_config = json.load(f)
+
+		if packaging.version.parse(env_config.get('version')) < packaging.version.parse('0.10.0'):
+			raise click.ClickException('The target environment version is lower than the supported minimum. Please upgrade the environment first.')
+
+		if env_config.get('namespace', 'default').lower() == namespace:
+			raise click.ClickException('The "%s" namespace is already in use in the target environment. Please use a different namespace.' % namespace)
+
+		if not env_config.get('provider_token'):
+			raise click.ClickException('Could not find a provider token in the target environment.')
+
+		# Force the same provider token
+		provider_token = env_config.get('provider_token')
 
 	if not provider_token:
 		provider_token = util.get_sail_default('provider-token')
@@ -65,7 +99,6 @@ def init(ctx, provider_token, email, size, region, force):
 	app_id = response['app_id']
 	secret = response['secret']
 	hostname = response['hostname']
-	namespace = 'default' # TODO: Add support for namespaces
 
 	click.echo('- Claimed application id: %s' % app_id)
 
@@ -84,6 +117,13 @@ def init(ctx, provider_token, email, size, region, force):
 		'namespace': namespace,
 		'version': __version__,
 	}
+
+	if environment:
+		config['profile_key'] = env_config['profile_key']
+		config['ip'] = env_config['ip']
+		config['droplet_id'] = env_config['droplet_id']
+		config['key_id'] = env_config['key_id']
+
 	util.update_config(config)
 
 	passwords = {}
@@ -91,9 +131,14 @@ def init(ctx, provider_token, email, size, region, force):
 		passwords[key] = ''.join(secrets.choice(string.ascii_letters
 			+ string.digits) for i in range(48))
 
-	# Provision an environment
-	_provision(provider_token, size, region)
-	_configure()
+	# Provision a new environment or copy existing
+	if not environment:
+		_provision(provider_token, size, region)
+		_configure()
+	else:
+		_copy_environment(environment)
+
+	# Run the installs
 	_install(passwords)
 
 	# Download files from production
@@ -126,6 +171,51 @@ def init(ctx, provider_token, email, size, region, force):
 
 	click.echo()
 	click.echo('For support and documentation visit sailed.io')
+
+def _copy_environment(environment):
+	root = util.find_root()
+	config = util.config()
+
+	click.echo('- Copying environment')
+
+	shutil.copy(environment / 'ssh.key', '%s/.sail/ssh.key' % root)
+	os.chmod('%s/.sail/ssh.key' % root, 0o600)
+	shutil.copy(environment / 'ssh.key.pub', '%s/.sail/ssh.key.pub' % root)
+	os.chmod('%s/.sail/ssh.key.pub' % root, 0o644)
+
+	click.echo('- Requesting DNS record')
+	response = util.request('/ip/', json={
+		'ip': config['ip'],
+	})
+
+	click.echo('- Checking SSH connection')
+	c = util.connection()
+
+	try:
+		c.run('hostname')
+	except Exception as e:
+		raise e
+		raise click.ClickException('Could not connect via SSH. Make sure the environment is healthy.')
+
+	try:
+		remote_config = json.loads(c.run('cat /etc/sail/config.json').stdout)
+	except:
+		raise click.ClickException('Could not read /etc/sail/config.json')
+
+	if not remote_config.get('namespaces'):
+		raise click.ClickException('Could not find any namespaces data in the remote config. This environment looks broken.')
+
+	# Make sure our new namespace is unique
+	if config['namespace'] in remote_config['namespaces']:
+		raise click.ClickException('The "%s" namespace already exists in this environment.' % config['namespace'])
+
+	remote_config['namespaces'].append(config['namespace'])
+	c.put(io.StringIO(json.dumps(remote_config)), '/etc/sail/config.json')
+
+	click.echo('- Writing server keys to .sail/known_hosts')
+	with open('%s/.sail/known_hosts' % root, 'w+') as f:
+		r = subprocess.run(['ssh-keyscan', '-t', 'rsa,ecdsa', '-H', config['ip'],
+			config['hostname']], stdout=f, stderr=subprocess.DEVNULL)
 
 def _provision(provider_token, size, region):
 	root = util.find_root()
@@ -251,7 +341,8 @@ def _configure():
 	# Generate an /etc/sail/config.json
 	config_json = {
 		'app_id': config['app_id'],
-		'profile_key': config['profile_key']
+		'profile_key': config['profile_key'],
+		'namespaces': [config['namespace']],
 	}
 
 	c.run('mkdir /etc/sail')
@@ -269,16 +360,23 @@ def _configure():
 	click.echo('- Waiting for cloud-init to complete')
 	util.wait(wait_for_cloud_init, timeout=300, interval=10)
 
-	c.run('mkdir -p /etc/nginx/conf.d/shared')
+	c.run('mkdir -p /etc/nginx/conf.d/extras')
 	c.put(sail.TEMPLATES_PATH + '/nginx.main.conf', '/etc/nginx/nginx.conf')
-	c.put(sail.TEMPLATES_PATH + '/nginx.shared.conf', '/etc/nginx/conf.d/shared/sail.conf')
+	c.put(sail.TEMPLATES_PATH + '/nginx.shared.conf', '/etc/nginx/conf.d/extras/sail.conf')
+	c.put(sail.TEMPLATES_PATH + '/nginx.certbot.conf', '/etc/nginx/conf.d/extras/certbot.conf')
 	c.put(sail.TEMPLATES_PATH + '/prepend.php', '/etc/sail/prepend.php')
 	c.put(sail.TEMPLATES_PATH + '/php.ini', '/etc/php/7.4/fpm/php.ini')
 	c.put(sail.TEMPLATES_PATH + '/php.ini', '/etc/php/7.4/cli/php.ini')
 
+	# Make sure certbot.conf is in action.
+	c.run('systemctl reload nginx')
+
 	# Get xhprof
 	c.run('curl -L https://github.com/kovshenin/xhprof/releases/download/0.10.0-sail/xhprof.so.gz -o /tmp/xhprof.so.gz')
 	c.run('gunzip /tmp/xhprof.so.gz && mv /tmp/xhprof.so /usr/lib/php/20190902/xhprof.so')
+
+	# Get dhparams
+	c.run('curl -L https://raw.githubusercontent.com/certbot/certbot/4abd81e2186eddc67551d61a8260440bd177d18d/certbot/certbot/ssl-dhparams.pem -o /etc/nginx/conf.d/extras/ssl-dhparams.pem')
 
 	# Permissions
 	c.run('chown -R www-data. /var/www')
@@ -287,19 +385,15 @@ def _install(passwords):
 	config = util.config()
 	c = util.connection()
 
-	# Generate default server config and install cert.
-	click.echo('- Installing default SSL certificate')
-	c.put(io.StringIO(util.template('nginx.server.conf', {
-		'server_names': [config['hostname']],
-		'root': util.remote_path('/public'),
-	})),'/etc/nginx/conf.d/%s.conf' % config['hostname'])
+	# Request an SSL cert
+	click.echo('- Requesting SSL certificate')
 
 	retry = 3
 	https = False
 	while retry:
 		retry -= 1
 		try:
-			c.run('certbot -n --register-unsafely-without-email --agree-tos --nginx --redirect -d %s' % config['hostname'])
+			c.run('certbot -n certonly --register-unsafely-without-email --agree-tos --standalone --http-01-port 8088 -d %s' % config['hostname'])
 			https = True
 			break
 		except:
@@ -317,6 +411,9 @@ def _install(passwords):
 		'primary': True
 	})
 	util.update_config(config)
+
+	# Update nginx configs with cert data.
+	domains._update_nginx_config()
 
 	# Prepare release directories
 	remote_path = util.remote_path()
@@ -367,19 +464,64 @@ def _install(passwords):
 	c.run('systemctl reload php7.4-fpm.service')
 
 @cli.command()
-@click.option('--yes', '-y', is_flag=True, help='Force Y on overwriting local copy')
-def destroy(yes):
-	'''Shutdown and destroy the production droplet'''
+@click.option('--yes', '-y', is_flag=True, help='Force yes on the are-you-sure prompt')
+@click.option('--environment', is_flag=True, help='Force destroy environment, even if other namespaces exist')
+def destroy(yes, environment):
+	'''Destroy an application namespace and/or the environment'''
 	root = util.find_root()
 	config = util.config()
 
 	if not yes:
-		click.confirm('All droplet data will be scrubbed and irretrievable. Are you sure?', abort=True)
+		click.confirm('All application data will be scrubbed and irretrievable. Are you sure?', abort=True)
 
 	click.echo('# Destroying application')
 
+	namespaces = []
+	try:
+		c = util.connection()
+		remote_config = json.loads(c.run('cat /etc/sail/config.json').stdout)
+		namespaces = remote_config.get('namespaces', [])
+	except:
+		pass
+
+	# It's the only namespace, destoy the environment
+	if len(namespaces) < 1 or namespaces == [config['namespace']]:
+		environment = True
+	elif not environment and config['namespace'] == 'default':
+		raise click.ClickException('You asked to destroy the default namespace, but other namespaces exist in this environment. Use --environment to destroy the entire environment.')
+
 	data = util.request('/destroy/', method='DELETE')
 
+	if environment:
+		_destroy_environment()
+	else:
+		_destroy_namespace()
+
+	click.echo('- Removing .sail/*')
+	shutil.rmtree(root + '/.sail')
+
+def _destroy_namespace():
+	config = util.config()
+	c = util.connection()
+
+	namespace = config['namespace']
+
+	c.run('rm -rf /var/www/_%s' % namespace)
+	c.run('rm -rf /etc/nginx/conf.d/%s.conf' % namespace)
+	c.run('mysql -e "DROP DATABASE \\`wordpress_%s\\`;"' % namespace)
+	c.run('mysql -e "DROP USER \\`wordpress_%s\\`@localhost"' % namespace)
+
+	# Update the remote config.
+	remote_config = json.loads(c.run('cat /etc/sail/config.json').stdout)
+	remote_config['namespaces'].remove(config['namespace'])
+	c.put(io.StringIO(json.dumps(remote_config)), '/etc/sail/config.json')
+
+	# TODO: Maybe revoke/delete SSL certificates for this namespace
+
+	c.run('systemctl reload nginx')
+
+def _destroy_environment():
+	config = util.config()
 	try:
 		droplet = digitalocean.Droplet(
 			token=config['provider_token'],
@@ -399,9 +541,6 @@ def destroy(yes):
 		click.echo('- SSH key destroyed successfully')
 	except:
 		click.echo('- Error destroying SSH key')
-
-	click.echo('- Removing .sail/*')
-	shutil.rmtree(root + '/.sail')
 
 @cli.command()
 @click.option('--provider-token', help='Your DigitalOcean API token, must be read-write. You can set a default token with: sail config provider-token <token>')