+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+* Added: Namespaces and environments to run multiple applications in the same environment. Use `--namespace` and `--environment` with `init`.
 * Added: New `pre-deploy` hooks in .sail, these will run every time deploy is invoked
 * Added: New `sail diff` command, shortcut for `sail deploy --dry-run` and `sail download --dry-run`
 * Changed: New `--skip-hooks` or `--no-verify` options to `sail deploy` to skip running hooks